requirements.txt
configparser == 6.0.0
customtkinter == 5.2.1
darkdetect == 0.8.0
openpyxl == 3.1.2
packaging == 23.2
pandas == 2.1.1
pillow == 10.2.0
progress == 1.6
pytz == 2023.3.post1
requests == 2.31.0
xlwings == 0.30.12